refactor(pipe_capacity): relocate DNV-OS-F201 LRFD/WSD check into pipe_capacity/custom (epic llm-wiki#767)
Moves code_checks/dnv_os_f201_lrfd_wsd.py -> structural/pipe_capacity/custom/DNV_OS_F201/
to sit alongside DNV_OS_F101 / API_RP_2RD / API_STD_2RD. Test moved to
tests/structural/pipe_capacity/ + import updated. 9 tests pass (namespace-package import).
